+++ tfjs-core/src/ops/ops.ts
@@ -90,6 +90,7 @@ export {pad4d} from './pad4d';
 export {pool} from './pool';
 export {pow} from './pow';
 export {prelu} from './prelu';
+export {prod} from './prod';
 export {rand} from './rand';
 export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
