+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -71,6 +71,7 @@ const CHAINED_OPS = [
   'pool',
   'pow',
   'prelu',
+  'prod',
   'relu',
   'resizeBilinear',
   'resizeNearestNeighbor',
